Move FCPXML import to assets panel
